% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewascatalog.R
\name{ewascatalog}
\alias{ewascatalog}
\title{ewascatalog}
\usage{
ewascatalog(
  query,
  type = c("cpg", "loc", "region", "gene", "trait", "efo", "study"),
  url = "http://www.ewascatalog.org"
)
}
\arguments{
\item{query}{Query text.}

\item{type}{Type of query, either 'cpg', 'region', 'gene', 'trait', 'efo', 'study' (Default: 'cpg').}
}
\value{
Data frame of EWAS Catalog results.
}
\description{
ewascatalog queries the EWAS Catalog from R.
}
\examples{
# CpG
res <- ewascatalog("cg00029284","cpg")

# Region
res <- ewascatalog("6:15000000-25000000","region")

# Gene
res <- ewascatalog("FTO","gene")

# Trait
res <- ewascatalog("body mass index","trait")

# EFO
res <- ewascatalog("EFO_0002950","efo")

# Study
res <- ewascatalog("27040690","study")

}
\author{
James R Staley <js16174@bristol.ac.uk>
}
