% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewascatalog.R
\name{get_meta_data}
\alias{get_meta_data}
\title{get_meta_data}
\usage{
get_meta_data()
}
\value{
Tibble of all EWAS Catalog meta-data
}
\description{
get_meta_data extracts all the meta-data from The EWAS Catalog
}
\examples{

studies <- get_meta_data()
}
\author{
Thomas Battram <thomas.battram@bristol.ac.uk>
}
