% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewascatalog.R
\name{ewascatalog}
\alias{ewascatalog}
\title{ewascatalog}
\usage{
ewascatalog(cpgquery = NULL, regionquery = NULL, genequery = NULL,
  traitquery = NULL)
}
\arguments{
\item{cpgquery}{a vector of CpG sites.}

\item{regionquery}{a vector of genomic regions.}

\item{genequery}{a vector of gene names.}

\item{traitquery}{a vector of trait names.}
}
\value{
Data frame of EWAS Catalog results.
}
\description{
ewascatalog queries the EWAS Catalog from R.
}
\examples{
# CpG
res <- ewascatalog(cpgquery="cg00029284")

# Region
res <- ewascatalog(regionquery="6:15000000-25000000")

# Gene
res <- ewascatalog(genequery="FTO")

# Trait
res <- ewascatalog(traitquery="body mass index")
}
\author{
James R Staley <jrstaley95@gmail.com>
}
